Initialising the FramesPerSecond class int DXSystem and injecting it into the Graphics class so that FPS can be rendered as text in the application

+++ Intellum/engine/system_metrics/FramesPerSecond.h
@@ -9,7 +9,7 @@ class FramesPerSecond
 private:
 	int _framesPerSecond;
 	int _frameCount;
-	int _deltaCount;
+	float _deltaCount;
 
 public:
 	FramesPerSecond();
